// src/pages/ExamsForm/types/ExamFields.ts
export type ExamFields = {
  id_medical_record?: string;
  anormalidades_encontradas: string;
  higiene_bucal: string;
  exames_complementares: string;
  exame_intrabucal_labios: string;
  exame_intrabucal_palato_duro: string;
  exame_intrabucal_lingua: string;
  exame_intrabucal_faringe: string;
  exame_intrabucal_gengivas: string;
  exame_intrabucal_habitos: string;
  exame_intrabucal_mucosa_jugal: string;
  exame_intrabucal_palato_mole: string;
  exame_intrabucal_assoalho_bucal: string;
  exame_intrabucal_glan_salivares: string;
  exame_intrabucal_atm: string;
  exame_intrabucal_outros: string;
  resultados_encontrados: string;
};